+++ src/app/modals/terms/terms-txt.ts
