+++ src/app/modals/terms/terms.service.spec.ts
@@ -2,8 +2,8 @@ import { TestBed, inject } from '@angular/core/testing';
 import { MatDialogModule } from '@angular/material';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
-import { CoreUiModule } from '../../core-ui.module';
-import { ModalsModule } from '../../../modals/modals.module';
+import { CoreUiModule } from '../../core-ui/core-ui.module';
+import { ModalsModule } from './../modals.module';
 import { TermsService } from './terms.service';
 
 describe('TermsService', () => {
