+++ tslint.json
@@ -22,8 +22,7 @@
     "label-position": true,
     "max-line-length": [
       true,
-      140,
-      "ignoreUrls"
+      140
     ],
     "member-access": false,
     "member-ordering": [
@@ -117,6 +116,11 @@
     "directive-class-suffix": [true, "Directive"],
     "templates-use-public": true,
     "no-access-missing-member": true,
-    "invoke-injectable": true
+    "invoke-injectable": true,
+    "linterOptions": {
+      "exclude": [
+        "src/app/modals/terms/terms-txt.ts"
+      ]
+    }
   }
 }
